requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==5.0.6
django-admin==2.0.2
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.15.2
excel-base==1.0.4
isoweek==1.3.3
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
python-dateutil==2.9.0.post0
screen==1.0.1
six==1.16.0
sqlparse==0.5.0
TimeConvert==3.0.13
tomli==2.0.1
typing_extensions==4.12.2
tzlocal==5.2
xlwt==1.3.0
